redpos finished and redneg alliance working

+++ include/arm.h
@@ -3,7 +3,7 @@
 #ifndef ARM_H
 #define ARM_H
 
-void move_arm_to(float target, float time_limit);
+void move_arm_to(float target, int time_limit);
 void move_arm_up();
 void move_arm_down();
 void stop_arm();
